Add ocontourc, a version of contourc which yields oriented contours.
The default Matlab contour algorithm (contourc) does not orient the
contour lines in a consistent way.  I.e. if you walk along the line
from start to end, values larger than the isovalue may be on the left
or right hand side of the line.

This meant that the hatches drawn by hatchedcontours could be randomly
oriented -- which is clearly non ideal.