Fix bug with default axes and huge data.
Hatchedline prefers to not be the first thing plotted in a set of axes.
This is because hatchedline calls 'axes' to get the aspect ratio and
scale of the current plot axes.

If the default axes ([0 1 0 1]) are a bad match for your data, you will
likely get a blank plot or cause a terrible performance problem / crash.

This fix avoids the worst of the problem, but for best results, you
should still set up your axes before calling hatchedline.

Also introduces bug_1.m to demonstrate the fix.